requirements.txt
python-telegram-bot==20.0
python-dotenv==0.21.0
selenium==4.9.0
webdriver-manager==3.8.6